database factory

+++ database/migrations/2023_05_09_012746_create_orders_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('orders', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('buyer_id');
-            $table->decimal('total_amount', 10, 2);
+            $table->float('total_amount', 10, 2);
             $table->string('status'); // pending, shipped, delivered
             $table->timestamps();
